# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Test API Server"
        }
    ],
    "paths": {
        "/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get list of employees",
                "description": "Returns list of employees",
                "operationId": "getEmployeesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Store new employee",
                "description": "Returns employee data",
                "operationId": "storeEmployee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get employee information",
                "description": "Returns employee data",
                "operationId": "getEmployeeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update existing employee",
                "description": "Returns updated employee data",
                "operationId": "updateEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete existing employee",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreEmployeeRequest": {
                "title": "StoreEmployeeRequest",
                "description": "StoreEmployeeRequest",
                "required": [
                    "name",
                    "status"
                ],
                "properties": {
                    "name": {
                        "title": "Имя",
                        "description": "Имя сотрудника",
                        "type": "string",
                        "example": "New Employee"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email сотрудника",
                        "type": "string",
                        "format": "email",
                        "example": "damien@example.net"
                    },
                    "status": {
                        "title": "Статус",
                        "description": "Статус сотрудника",
                        "type": "string",
                        "example": "Работает"
                    }
                },
                "type": "object"
            },
...
            "UpdateEmployeeRequest": {
                "title": "UpdateEmployeeRequest",
                "description": "UpdateEmployeeRequest",
                "required": [
                    "name",
                    "status"
                ],
                "properties": {
                    "name": {
                        "title": "Имя",
                        "description": "Имя сотрудника",
                        "type": "string",
                        "example": "Updated Employee"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email сотрудника",
                        "type": "string",
                        "format": "email",
                        "example": "damien@example2.net"
                    },
                    "status": {
                        "title": "Статус",
                        "description": "Статус сотрудника",
                        "type": "string",
                        "example": "В отпуске"
                    }
                },
                "type": "object"
            },
...
            "EmployeeResource": {
                "title": "EmployeeResource",
                "description": "Employee resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Employee"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EmployeeResource"
                }
            },
...
            "Employee": {
                "title": "Сотрудник",
                "description": "Employee model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Имя",
                        "description": "Имя сотрудника",
                        "type": "string",
                        "example": "Cayla Kling"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email сотрудника",
                        "type": "string",
                        "format": "email",
                        "example": "kiehn.damien@example.net"
                    },
                    "status": {
                        "title": "Статус",
                        "description": "Статус сотрудника",
                        "type": "string",
                        "example": "Работает"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Employee"
                }
            },
...
        }
    },
...
}